+++ paymentservice/src/main/java/com/techworld/paymentservice/model/PaymentResponse.java
@@ -1,11 +1,11 @@
 package com.techworld.paymentservice.model;
 
-import java.time.Instant;
-
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
+
+import java.time.Instant;
 
 @Data
 @NoArgsConstructor
@@ -16,7 +16,7 @@ public class PaymentResponse {
     private long paymentId;
     private String status;
     private PaymentMode paymentMode;
-    private long amount;
+    private double amount;
     private Instant paymentDate;
     private long bookingId;
 }
